+++ Lab6_BankAccount/Lab6_BankAccount/Models/Account.cs
@@ -12,10 +12,9 @@ namespace Lab6_BankAccount.Models
     using System;
     using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
-
     public partial class Account
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         [DisplayFormat(DataFormatString = "{0:MM/dd/yyyy}")]
